Create a RankingHistoryContext
