+++ frontend/src/components/RankingsHistoryComponent.client.tsx
@@ -1,10 +1,4 @@
 import React, { useEffect, useState } from 'react';
-import axios from 'axios';
-import {
-  PlayerHistory,
-  GraphData,
-  PlayerHistoryEntry,
-} from '@/types/playerHistoryTypes';
 import {
   LineChart,
   Line,
@@ -15,33 +9,16 @@ import {
   Legend,
   ResponsiveContainer,
 } from 'recharts';
-import { parseISO, subDays, format } from 'date-fns';
 import { capitalizeFirstLetter } from '@/utils/string';
 import { usePlayerContext } from '@/contexts/PlayerContext';
+import { useRankingHistoryContext } from '@/contexts/RankingHistoryContext';
 
 const RankingsHistoryComponent = () => {
   const { players } = usePlayerContext();
-  const [data, setData] = useState<GraphData[]>([]);
+  const { data, loading, error } = useRankingHistoryContext();
   const [playerColors, setPlayerColors] = useState<{ [key: string]: string }>(
     {},
   );
-
-  useEffect(() => {
-    const fetchHistoryData = async () => {
-      try {
-        const response = await axios.get<PlayerHistory[]>(
-          'https://brs.aragorn-media-server.duckdns.org/players/history?type=RANK',
-        );
-        const preprocessedData = preprocessDataToAddOldEntries(response.data);
-        const transformedData = transformDataForGraph(preprocessedData);
-        setData(transformedData);
-      } catch (error) {
-        console.error('There was an error fetching the history data:', error);
-      }
-    };
-
-    fetchHistoryData();
-  }, []);
 
   useEffect(() => {
     const colorsMapping: { [key: string]: string } = {};
@@ -50,57 +27,6 @@ const RankingsHistoryComponent = () => {
     });
     setPlayerColors(colorsMapping);
   }, [players]);
-
-  const preprocessDataToAddOldEntries = (
-    players: PlayerHistory[],
-  ): PlayerHistory[] => {
-    return players.map((player) => {
-      if (player.history.length > 0) {
-        const sortedHistory = player.history.sort(
-          (a, b) => parseISO(a.date).getTime() - parseISO(b.date).getTime(),
-        );
-        const mostRecentEntry = sortedHistory[0];
-
-        const newDate = subDays(parseISO(mostRecentEntry.date), 7);
-
-        const newEntry: PlayerHistoryEntry = {
-          oldRank: mostRecentEntry.oldRank,
-          newRank: mostRecentEntry.oldRank,
-          date: format(newDate, 'yyyy-MM-dd'),
-        };
-
-        return {
-          ...player,
-          history: [...player.history, newEntry],
-        };
-      }
-      return player;
-    });
-  };
-
-  const transformDataForGraph = (players: PlayerHistory[]): GraphData[] => {
-    let graphData: GraphData[] = [];
-    players.forEach((player) => {
-      player.history.forEach((entry) => {
-        let existingDateEntry = graphData.find((e) => e.date === entry.date);
-        const playerNameCapitalized = capitalizeFirstLetter(player.playerName);
-        if (existingDateEntry) {
-          existingDateEntry[playerNameCapitalized] = entry.newRank;
-        } else {
-          const newEntry: GraphData = {
-            date: entry.date,
-            [playerNameCapitalized]: entry.newRank,
-          };
-          graphData.push(newEntry);
-        }
-      });
-    });
-
-    graphData.sort(
-      (a, b) => parseISO(a.date).getTime() - parseISO(b.date).getTime(),
-    );
-    return graphData;
-  };
 
   return (
     <ResponsiveContainer width='100%' height={400}>
