+++ frontend/src/pages/_app.tsx
@@ -2,32 +2,18 @@ import '../app/globals.css';
 import type { AppProps } from 'next/app';
 import LayoutComponent from '@/components/LayoutComponent';
 import NavigationComponent from '@/components/NavigationComponent';
-import React, { useEffect, useState } from 'react';
-import axios from 'axios';
-import { Player } from '@/types/player';
+import React from 'react';
 import { PlayerProvider } from '@/contexts/PlayerContext';
+import { RankingHistoryProvider } from '@/contexts/RankingHistoryContext';
 
 function MyApp({ Component, pageProps }: AppProps) {
-  const [players, setPlayers] = useState<Player[]>([]);
-
-  useEffect(() => {
-    const fetchPlayers = async () => {
-      const response = await axios.get(
-        'https://brs.aragorn-media-server.duckdns.org/players',
-      );
-      const sortedData = response.data.sort(
-        (a: Player, b: Player) => a.playerRank - b.playerRank,
-      );
-      setPlayers(sortedData);
-    };
-    fetchPlayers();
-  }, []);
-
   return (
     <LayoutComponent>
       <PlayerProvider>
-        <NavigationComponent />
-        <Component {...pageProps} />
+        <RankingHistoryProvider>
+          <NavigationComponent />
+          <Component {...pageProps} />
+        </RankingHistoryProvider>
       </PlayerProvider>
     </LayoutComponent>
   );
