+++ frontend/src/utils/string.ts
@@ -1,6 +1,6 @@
 export const capitalizeFirstLetter = (string: string) => {
   return string
     .split(' ')
-    .map(word => word.charAt(0).toUpperCase() + word.slice(1))
+    .map((word) => word.charAt(0).toUpperCase() + word.slice(1))
     .join(' ');
 };
